+++ src/displayEmailList/displayEmailCard/DisplayEmailCard.js
@@ -1,0 +1,14 @@
+import './DisplayEmailCard.css'
+
+function displayEmailCard() {
+
+    return (
+        <div>
+            <h2 className="contactName">Maria Smith</h2>
+            <p className="emailSubject"></p>
+            <p className="emailContent"></p>
+        </div>
+    )
+}
+
+export default DisplayEmailCard
