--- src/displayEmailSnippet/displayEmailCard/DisplayEmailCard.js
@@ -1,14 +1,0 @@
-import './DisplayEmailCard.css'
-
-function displayEmailCard() {
-
-    return (
-        <div>
-            <h2 className="contactName">Maria Smith</h2>
-            <p className="emailSubject"></p>
-            <p className="emailContent"></p>
-        </div>
-    )
-}
-
-export default DisplayEmailCard
